python-ml-api/requirements.txt
tensorflow==2.19.0
numpy==1.26.0
opencv-python==4.8.1.78
pillow==11.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6